refactor: убрать DirectorDbStorage из FilmDbStorage
