+++ src/main/java/ru/yandex/practicum/filmorate/storage/director/DirectorDbStorage.java
@@ -1,6 +1,6 @@
 package ru.yandex.practicum.filmorate.storage.director;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Primary;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
@@ -11,23 +11,19 @@ import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.exceptions.InvalidIdException;
 import ru.yandex.practicum.filmorate.mapper.Mapper;
 import ru.yandex.practicum.filmorate.model.Director;
+import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.model.Genre;
 
-import java.util.Collection;
-import java.util.Objects;
-import java.util.Optional;
+import java.util.*;
 
 import static ru.yandex.practicum.filmorate.constants.SqlQueryConstantsForFilm.*;
 import static ru.yandex.practicum.filmorate.constants.SqlQueryConstantsForUser.SQL_QUERY_REMOVE_DIRECTOR;
 
 @Component
 @Primary
+@RequiredArgsConstructor
 public class DirectorDbStorage implements DirectorStorage {
     private final NamedParameterJdbcTemplate namedParameterJdbcTemplate;
-
-    @Autowired
-    public DirectorDbStorage(NamedParameterJdbcTemplate namedParameterJdbcTemplate) {
-        this.namedParameterJdbcTemplate = namedParameterJdbcTemplate;
-    }
 
     @Override
     public Collection<Director> findAllDirectors() {
@@ -80,4 +76,37 @@ public class DirectorDbStorage implements DirectorStorage {
         namedParameterJdbcTemplate.update(SQL_QUERY_REMOVE_DIRECTOR, parameters);
     }
 
+    @Override
+    public List<Film> findFilmsByDirectorAndSort(int directorId, String query) {
+        findDirectorById(directorId)
+                .orElseThrow(() -> new InvalidIdException("Нет режиссера с id " + directorId));
+        ArrayList<Film> films = new ArrayList<>();
+        SqlRowSet filmRows =
+                namedParameterJdbcTemplate.getJdbcTemplate().queryForRowSet(query,
+                        directorId);
+        while (filmRows.next()) {
+            films.add(addGenreAndDirectorToFilm(Mapper.makeFilm(filmRows)));
+        }
+        return films;
+    }
+
+    private Film addGenreAndDirectorToFilm(Film film) {
+        SqlRowSet genreRows = namedParameterJdbcTemplate
+                .getJdbcTemplate().queryForRowSet(SQL_QUERY_TAKE_FILMS_GENRE_AND_DIRECTOR_BY_ID, film.getId());
+        Set<Genre> genres = new TreeSet<>(Comparator.comparingLong(Genre::getId));
+        while (genreRows.next()) {
+            int genreId = genreRows.getInt("genre_id");
+            String genreName = genreRows.getString("genre_name");
+            int directorId = genreRows.getInt("director_id");
+            String directorName = genreRows.getString("director_name");
+            if (genreId != 0 && genreName != null) {
+                genres.add(new Genre(genreId, genreName));
+            }
+            if (directorId != 0 && directorName != null) {
+                film.addDirectorToFilm(new Director(directorId, directorName));
+            }
+        }
+        film.setGenres(genres);
+        return film;
+    }
 }
