+++ src/main/java/ru/yandex/practicum/filmorate/storage/director/DirectorStorage.java
@@ -1,8 +1,10 @@
 package ru.yandex.practicum.filmorate.storage.director;
 
 import ru.yandex.practicum.filmorate.model.Director;
+import ru.yandex.practicum.filmorate.model.Film;
 
 import java.util.Collection;
+import java.util.List;
 import java.util.Optional;
 
 public interface DirectorStorage {
@@ -17,4 +19,5 @@ public interface DirectorStorage {
 
     void removeDirector(long directorId);
 
+    List<Film> findFilmsByDirectorAndSort(int directorId, String query);
 }
